# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ProofLoop API
  version: 0.2.0
  description: |
    Evidence-first verification API.
    When `PROOFLOOP_API_TOKEN` is set, mutating routes require `Authorization: Bearer <token>`.
servers:
  - url: http://localhost:8787
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /api/github/webhook:
    post:
      summary: GitHub webhook (PR opened/synchronize/reopened)
      description: |
        Uses GitHub App installation token when `GITHUB_APP_ID` + `GITHUB_APP_PRIVATE_KEY` are set,
        otherwise `GITHUB_TOKEN`. Can clone/fetch into `PROOFLOOP_WORKSPACE_ROOT` when repo has no localPath.
      responses:
        '200':
          description: Accepted
        '401':
          description: Invalid signature
        '503':
          description: Webhook secret required in production
  /api/gitlab/webhook:
    post:
      summary: GitLab webhook (MR open/update/reopen)
      description: |
        Verifies `X-Gitlab-Token` against `GITLAB_WEBHOOK_SECRET`. Uses `GITLAB_TOKEN` + `GITLAB_URL`
        for commit status, MR notes, and optional checkout into `PROOFLOOP_WORKSPACE_ROOT`.
      responses:
        '200':
          description: Accepted
        '401':
          description: Invalid token
        '503':
          description: Webhook secret required in production